+++ Source/CesiumRuntime/Private/CesiumGltfPointsComponent.h
@@ -14,6 +14,17 @@ public:
   UCesiumGltfPointsComponent();
   virtual ~UCesiumGltfPointsComponent();
 
+  // Whether the tile that contains this point component uses additive
+  // refinement.
+  bool UsesAdditiveRefinement;
+
+  // The geometric error of the tile containing this point component.
+  float GeometricError;
+
+  // The dimensions of the point component. Used to estimate the geometric
+  // error.
+  glm::vec3 Dimensions;
+
   // Override UPrimitiveComponent interface.
   virtual FPrimitiveSceneProxy* CreateSceneProxy() override;
 };